+++ R/runMSE.R
@@ -23,7 +23,7 @@ Simulate <- function(OM=MSEtool::testOM, parallel=FALSE, silent=FALSE) {
   }
 
   # ---- Set up parallel processing ----
-  ncpus <- set_parallel(parallel)
+  ncpus <- set_parallel(any(parallel))
 
   set.seed(OM@seed) # set seed for reproducibility
   nsim <- OM@nsim # number of simulations
@@ -1081,6 +1081,14 @@ Project <- function (Hist=NULL, MPs=NA, parallel=FALSE, silent=FALSE,
   # ---- Setup ----
   if (class(Hist) !='Hist')
     stop('Must provide an object of class `Hist`')
+  
+  if (length(parallel) < length(MPs)) {
+    if (length(parallel) > 1) {
+      stop("length(parallel) must be equal to length(MPs)")
+    } else {
+      parallel <- rep(parallel, length(MPs))
+    }
+  }
 
   OM <- Hist@OM
   set.seed(OM@seed) # set seed for reproducibility
@@ -1102,9 +1110,9 @@ Project <- function (Hist=NULL, MPs=NA, parallel=FALSE, silent=FALSE,
   if (nMP < 1) stop("No valid MPs found", call.=FALSE)
 
   isrunning <- snowfall::sfIsRunning()
-  if (!parallel & isrunning) snowfall::sfStop()
+  if (all(!parallel) & isrunning) snowfall::sfStop()
 
-  if (parallel) {
+  if (any(parallel)) {
     if (!isrunning) setup()
     Export_customMPs(MPs)
   }
@@ -1219,7 +1227,7 @@ Project <- function (Hist=NULL, MPs=NA, parallel=FALSE, silent=FALSE,
     Data_MP <- MSElist[[mm]]
     Data_MP@Misc <- Data_Misc # add StockPars etc back to Data object
 
-    if(!silent) message(mm, "/", nMP, " Running MSE for ", MPs[mm])
+    if(!silent) message(mm, "/", nMP, " Running MSE for ", MPs[mm], ifelse(parallel[mm], " in parallel", ""))
     checkNA <- rep(0, OM@proyears) # save number of NAs
     # years management is updated
     upyrs <- seq(from=1, to=proyears, by=interval[mm])
@@ -1312,7 +1320,7 @@ Project <- function (Hist=NULL, MPs=NA, parallel=FALSE, silent=FALSE,
 
     StockPars$N_P <- N_P
     # -- Apply MP in initial projection year ----
-    runMP <- applyMP(Data=Data_MP, MPs = MPs[mm], reps = reps, silent=TRUE)  # Apply MP
+    runMP <- applyMP(Data=Data_MP, MPs = MPs[mm], reps = reps, silent=TRUE, parallel = parallel[mm])  # Apply MP
 
     MPRecs <- runMP[[1]][[1]] # MP recommendations
     Data_p <- runMP[[2]] # Data object object with saved info from MP
@@ -1506,7 +1514,7 @@ Project <- function (Hist=NULL, MPs=NA, parallel=FALSE, silent=FALSE,
 
 
         # --- apply MP ----
-        runMP <- applyMP(Data=Data_MP, MPs = MPs[mm], reps = reps, silent=TRUE)  # Apply MP
+        runMP <- applyMP(Data=Data_MP, MPs = MPs[mm], reps = reps, silent=TRUE, parallel = parallel[mm])  # Apply MP
         MPRecs <- runMP[[1]][[1]] # MP recommendations
         Data_p <- runMP[[2]] # Data object object with saved info from MP
         Data_p@TAC <- MPRecs$TAC
@@ -1750,7 +1758,7 @@ Project <- function (Hist=NULL, MPs=NA, parallel=FALSE, silent=FALSE,
 #' @param Hist Should model stop after historical simulations? Returns an object of
 #' class 'Hist' containing all historical data
 #' @param silent Should messages be printed out to the console?
-#' @param parallel Logical. Should the MSE be run using parallel processing?
+#' @param parallel Logical. Should the MSE be run using parallel processing? Can be a vector of length(MPs)
 #' @param extended Logical. Return extended projection results?
 #' if TRUE, `MSE@Misc$extended` is a named list with extended data
 #' (including historical and projection by area), and extended version of `MSE@Hist`
@@ -1789,6 +1797,16 @@ runMSE <- function(OM=MSEtool::testOM, MPs = NA, Hist=FALSE, silent=FALSE,
   } else {
     stop("You must specify an operating model")
   }
+  
+  # check parallel
+  if (length(parallel) < length(MPs)) {
+    if (length(parallel) > 1) {
+      stop("length(parallel) must be equal to length(MPs)")
+    } else {
+      parallel <- rep(parallel, length(MPs))
+    }
+  }
+  
 
   # check MPs
   if (checkMPs & !Hist)